% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordermle.R
\name{ordermle}
\alias{ordermle}
\title{This provides MLEs for the parameters of a two-way ANCOVA model when the treatment effects of row factor are in simple order restrictions under heterogeneous variances}
\usage{
ordermle(Res, Cov, a, b, q)
}
\arguments{
\item{Res}{a real data set}

\item{Cov}{a real data set}

\item{a}{a positive integer}

\item{b}{a positive integer}

\item{q}{a positive integer}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
a=3;b=2;k=a*b;q=3
N1=c(10,10,10,15,15,15,10,10,10,15,15,15,10,10,10,15,15,15)
S=c(1,1,2,1,1,2,3,1,2,4,6,3,2,4,2,3,2,1)
g=NULL
for(i in 1:(k*q))
{
 g[[i]]=rnorm(N1[i],0,sqrt(S[i]))
}
Cov=g
G2=NULL
N=c(10,15,10,15,10,15);m=c(1,2,3,4,2,1)
for(i in 1:k)
{
G2[[i]]=rnorm(N[i],m[i],sqrt(S[i]))
}
Res=G2
ordermle(Res,Cov,a,b,q)
}
