% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TN.R
\name{TN}
\alias{TN}
\title{It gives the test statistic value and critical value of the test TN for testing homogeneity of treatment effects against ordered alternatives in two-way ANCOVA model}
\usage{
TN(Res, Cov, a, b, q, alpha)
}
\arguments{
\item{Res}{a real data set}

\item{Cov}{a real data set}

\item{a}{a positive integer}

\item{b}{a positive integer}

\item{q}{a positive integer}

\item{alpha}{a real number which lies in between 0 and  1}
}
\value{
numeric vector
}
\description{
More detailed description
}
\examples{
a=3;b=2;k=a*b;q=3;alpha=0.05
N1=c(10,10,10,15,15,15,10,10,10,15,15,15,10,10,10,15,15,15)
S=c(1,1,2,1,1,2,3,1,2,4,6,3,2,4,2,3,2,1)
g=NULL
for(i in 1:(k*q))
{
 g[[i]]=rnorm(N1[i],0,sqrt(S[i]))
}
Cov=g
G2=NULL
N=c(10,15,10,15,10,15);m=c(1,2,3,4,2,1)
for(i in 1:k)
{
G2[[i]]=rnorm(N[i],m[i],sqrt(S[i]))
}
Res=G2
TN(Res,Cov,a,b,q,alpha)
}
